+++ smart_media/fields.py
@@ -10,6 +10,11 @@ class SmartMediaField(FileField):
     A FileField which define the widget ``ClearableFileInputButton`` and default
     validator ``FileExtensionValidator`` for allowed extensions (as defined from
     setting ``SMARTIMAGE_ALLOWED_IMAGE_EXTENSIONS``).
+
+    .. Note::
+        Validation is pretty basic since it works on file extension, no malicious
+        Bitmap or SVG image files would be detected here. If it is a concern, you
+        will have to develop or use an additional validator.
     """
     widget = ClearableFileInputButton
     default_validators = [