+++ docs/history.rst
@@ -8,6 +8,8 @@ Version 0.3.0 - Unreleased
 --------------------------
 
 * Added missing link to Sorl documentation in tag docstring, close #5;
+* Fixed invisible thumb for SVG file in preview from ``FileInputButtonBase`` widget,
+  close #9;
 
 
 Version 0.2.2 - 2023/04/04